compiler: enable caching for -femit-h
As documented in #14416, when using `zig build-lib -femit-h` with
caching enabled, if the generated header file is removed, it will not
be generated again, unless the library source code is modified.

For consistent builds with `zig build`, assume that the file generated
by -femit using the default path is an artificact, and update
src/main.zig to enable caching.

Update the flushEmitH function in src/link/C.zig to use the
zig_cache_artifact_directory instead of the incorrect
local_cache_directory.

Updates #14416