Compilation: cap number of multi-package files reported, fix assertion failure
Errors caused by files existing in multiple packages will, for some
not-unreasonable package structures, report errors for a whole bunch of
files. Chances are, these all have the same root cause, so reporting
more than just the first few is pointless and just spams the output, so
let's cap the output at 5 errors of that type.

Also, if AstGen failed on a file marked multi-pkg for other reasons, we
were hitting an assertion failure because the ZIR wasn't loaded. To fix,
we just only look for nested includes if the ZIR is loaded.

Resolves: #14499