docgen: remove line support in printShell
Line support in printShell breaks the support to terminal colors, when a
color is active over multiple lines. An example is when printing
reference traces.

Remove line support in printShell, since they are not really necessary
like in printSourceBlock.

In genHtml, remove an empty line when printing the execution of the
executable generated by `zig build-exe`.

Update the "shell parsed" tests.
Remove the call to log.emerg, since it is no longer supported.
Add an extra space after "--build-option1 \", as documented.

Fixes #13280