std.crypto.hash.sha3: add TurboSHAKE
TurboSHAKE is a round-reduced version of SHAKE which is being
standardized by CFRG.

It offers the same practical security as SHA3 and SHAKE. Its security
builds up on the scrutiny that Keccak has received since its
publication.

Unlike SHA3 (but not unlike SHAKE), the output can be of any size.

TurboSHAKE also offers higher performance - It is nearly twice as
fast as SHAKE.

         sha3-256:        571 MiB/s
         sha3-512:        316 MiB/s
        shake-128:        723 MiB/s
   turboshake-128:       1334 MiB/s

Why it matters: TurboSHAKE is the fastest secure hash function we'll
have in Zig when compiling for the `baseline` CPU.

Since we already had SHAKE, TurboSHAKE is a trivial addition, and
it will immediately benefit from the further optimizations we'll
do to Keccak.