main: allow a relative --cache-dir to be special
Currently the Zig compiler driver allows a custom local cache directory
to be a relative path, resulting in the cache directory being created in
the working directory.

Update the buildOutputType and cmdBuild functions to make a relative
local cache path have the same semantic of the special "zig-cache"
directory, so that the cache directory will always be created in the
directory containing the build.zig file.