Build: allow a relative build --prefix to be special
Currently the build system allows a custom prefix path to be a relative
path, resulting in the prefix directory being created in the working
directory.

Update the resolveInstallPrefix method to make a relative prefix path
have the same semantic of the special "zig-out" directory, so that the
prefix directory will always be created in the directory containing the
build.zig file.