added an error for invalid pointer dereference
This error occurs when ' *. ' is typed instead of ' .* ' to dereference a pointer which is a common beginner error.

Sorry about the closed PR I tried fixing one thing and everything went wrong. This is also the first time I've ever made a PR to a large repo so there still might be problems with this one.