std.Build.RunStep: fix a bug when running a zig test
When running, as an example, a `zig build test` command in a build file,
what happens is the following:

  1. The builder_runner is executed with stdin closed, resulting in
     cwd_dir.fd to be 0 in std.Build.RunStep.
  2. The test is executed with stdin set to a pipe, so that spawn will
     fail, since cwd_dir.fs is now not a directory.

Add a new is_test_command field to RunStep; it must be set to true when
RunStep runs a zig test command.

Update the spawnChildAndCollect function, so that when running a zig
test command, stdin_behavior is set to .Ignore.

Fixes #15104