std.math.big.int: Implement depositBits and extractBits
This change implements depositBits and extractBits (equivalents of PDEP
and PEXT) for Zig's bit ints. This change lays the groundwork for
implementation of `@depositBits` and `@extractBits`.

Tests have been added to check the behaviour of these two functions.

The functions currently don't handle negative values (though negative
values may be converted to twos complement externally), and
aren't optimal in either memory or performance.