Revert "ci: let's not skip non native tests on macos!"
This reverts commit a34752c941c81e600c56670abb612fc86d0a2b73.

With #15277 merged, the macOS CI will still run the self-hosted x86_64
tests even with -Dskip-non-native. This was the main motivation for the
commit.

It would be nice to keep this flag enabled, however, we don't have
enough CI computing power, and the macOS CI lags behind with these extra
tests enabled.